// src/knx/enums/dpt.ts
export enum DPT {
    Switch = "1.001",
    Bool = "1.002",
    Enable = "1.003",
    UpDown = "1.008",
    OpenClose = "1.009",
    Reset = "1.015",
    Ack = "1.016",
    Trigger = "1.017",
    Occupancy = "1.018",
    Window_Door = "1.019",
    DayNight = "1.024",
  
    Value_Power = "14.056",
  
    Value_Temp = "9.001",
    Value_Humidity = "9.007",
    Value_AirQuality = "9.008",
  
    ActiveEnergy = "13.010",
  }
  